+++ NEWS.md
@@ -9,3 +9,5 @@
   exercises.
 - Added tracked radio-button and checkbox `learnr` question helpers.
 - Added tracked text and numeric `learnr` question helpers.
+- Added `tracked_question()` as a generic helper for built-in `learnr` question
+  types.